requirements.txt
web3
ruamel.yaml
git+https://github.com/ethereum/eth2.0-specs@dev
